create special folder for sevlets to use package
