+++ TWApp/src/main/java/savageTW/CheckServlet.java
@@ -1,11 +1,11 @@
+package savageTW;
+
 import com.google.gson.Gson;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.List;
 
 public class CheckServlet extends HttpServlet {
