+++ TWApp/src/main/java/savageTW/StatusServlet.java
@@ -1,11 +1,16 @@
+package savageTW;
+
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
-public class Test2Servlet extends HttpServlet {
+public class StatusServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        response.sendRedirect("/page");
+        PrintWriter pw = response.getWriter();
+
+        pw.println("<big style='color:red'>Application is running</big>");
     }
 }
