+++ TWApp/src/main/java/savageTW/Test2Servlet.java
@@ -1,14 +1,13 @@
+package savageTW;
+
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.PrintWriter;
 
-public class StatusServlet extends HttpServlet {
+public class Test2Servlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        PrintWriter pw = response.getWriter();
-
-        pw.println("<big style='color:red'>Application is running</big>");
+        response.sendRedirect("/page");
     }
 }
